requirements.txt
fastapi
uvicorn
groq
python-dotenv
pytest
pytest-cov
httpx
slowapi
